Whole-ship overview camera, click-to-place fixes, room decor, and EVA station-attach fix
- Shipwright dialogue now opens a whole-ship overview camera (zoom/pan via scroll
  wheel + right-drag) for choosing where to build a compartment
- Fixed room-visibility mask going black in the overview (BuildVisibilityMask
  exemption) and fixed click-to-place landing in the wrong spot (ScreenToShipLocal
  now accounts for scene zoom)
- Added procedural decor/textures for all 14 content-catalog room types
  (RoomDecor.Catalog.cs)
- Added a "give 100 credits" dev cheat button next to the enemy spawner
- Fixed EVA characters permanently freezing after pushing off near a station: a
  defensive anti-tunneling check in StepFreeFloating fired every tick once M59's
  static map parked ships flush against stations, since the exit point already read
  as "inside" the station on tick 0 - now only fires on a genuine outside->inside
  transition

Also includes in-progress work from elsewhere in the working tree (planet landing,
ship building/debris, celestial bodies, star systems, asteroid field rework).

+++ src/SpaceAdventure.Client/Game1.ShipEditor.Draw.cs
@@ -50,8 +50,8 @@ public partial class Game1
     private void DrawEditorRoom(CustomRoomDef room)
     {
         var rect = new Rectangle(
-            ShipEditorCanvas.X + room.X * ShipEditorCellSize, ShipEditorCanvas.Y + room.Y * ShipEditorCellSize,
-            room.Width * ShipEditorCellSize, room.Height * ShipEditorCellSize);
+            (int)(ShipEditorCanvas.X + room.X * ShipEditorCellSize), (int)(ShipEditorCanvas.Y + room.Y * ShipEditorCellSize),
+            (int)(room.Width * ShipEditorCellSize), (int)(room.Height * ShipEditorCellSize));
         _spriteBatch.Draw(_pixel, rect, new Color(46, 52, 66));
         DrawRectOutline(rect, new Color(120, 130, 150), 2f);
