+++ src/SpaceAdventure.Shared/Model/OreDeposit.cs
@@ -7,11 +7,13 @@ namespace SpaceAdventure.Shared.Model;
 //
 // The position is absolute and fixed - asteroids never move, so nothing here is expressed relative
 // to one - and Radius is what the flame has to touch, which is also what the client draws.
-public sealed record OreDeposit(string Id, string AsteroidId, float X, float Y, float MaxHp, float Radius = 0.55f)
+// X/Y are double, not float, for the same reason Asteroid's own doc comment gives - a deposit's
+// small offset from its parent asteroid's already-huge absolute position needs it too.
+public sealed record OreDeposit(string Id, string AsteroidId, double X, double Y, float MaxHp, float Radius = 0.55f)
 {
     public Vec2 Position => new(X, Y);
 
     // How far a point is from the block's body: 0 anywhere on or inside it. What the cutting flame
     // is tested against, segment sample by segment sample.
-    public float DistanceFrom(Vec2 point) => MathF.Max(0f, (point - Position).Length() - Radius);
+    public float DistanceFrom(Vec2 point) => MathF.Max(0f, (float)((point - Position).Length() - Radius));
 }
